+++ pdaltools/las_add_buffer.py
@@ -1,7 +1,7 @@
 import argparse
 import logging
 import os
-from typing import List
+from typing import Dict, List
 
 import pdal
 
@@ -46,6 +46,40 @@ def create_las_with_buffer(
     )
 
 
+def get_writer_parameters_from_reader_metadata(metadata: Dict, a_srs=None) -> Dict:
+    """As pdal las writers does not permit to pass easily metadata from one file as
+    parameters for a writer, use a trick to generate writer parameters from the
+    reader metadata of a previous pipeline:
+    This function uses the metadata from the reader of a pipeline to provide parameters
+    to pass to the writer of another pipeline
+
+    To be removed once https://github.com/PDAL/python/issues/147 is solved
+
+    Args:
+        metadata (Dict): metadata of an executed pipeline (that can be accessed using pipeline.metadata)
+    Returns:
+        Dict: parameters to pass to a pdal writer
+    """
+
+    reader_metadata = metadata["metadata"]["readers.las"]
+
+    params = {
+        "major_version": reader_metadata["major_version"],
+        "minor_version": reader_metadata["minor_version"],
+        "global_encoding": reader_metadata["global_encoding"],
+        "extra_dims": "all",
+        "scale_x": reader_metadata["scale_x"],
+        "scale_y": reader_metadata["scale_y"],
+        "scale_z": reader_metadata["scale_z"],
+        "offset_x": reader_metadata["offset_x"],
+        "offset_y": reader_metadata["offset_y"],
+        "offset_z": reader_metadata["offset_z"],
+        "dataformat_id": reader_metadata["dataformat_id"],
+        "a_srs": a_srs if a_srs else reader_metadata["comp_spatialreference"],
+    }
+    return params
+
+
 def las_merge_and_crop(
     input_dir: str,
     tile_filename: str,
@@ -77,12 +111,12 @@ def las_merge_and_crop(
                 (usually 1000m)
     """
     # List files to merge
-    Listfiles = create_list(input_dir, tile_filename, tile_width, tile_coord_scale)
+    files_to_merge = create_list(input_dir, tile_filename, tile_width, tile_coord_scale)
 
-    if len(Listfiles) > 0:
+    if len(files_to_merge) > 0:
         # Read and crop each file
         crops = []
-        for f in Listfiles:
+        for f in files_to_merge:
             pipeline = pdal.Pipeline()
             pipeline |= pdal.Reader.las(filename=f, override_srs=spatial_ref)
             pipeline |= pdal.Filter.crop(bounds=str(bounds))
@@ -92,13 +126,19 @@ def las_merge_and_crop(
             else:
                 crops.append(pipeline.arrays[0])
 
+            # Retrieve metadata before the pipeline is deleted
+            # As the last file of files_to_merge is the central one, metadata will contain the info
+            # from the central file after the last iteration of the for loop
+            metadata = pipeline.metadata
             del pipeline
+
+        params = get_writer_parameters_from_reader_metadata(metadata, a_srs=spatial_ref)
 
         # Merge
         pipeline = pdal.Filter.merge().pipeline(*crops)
 
         # Write
-        pipeline |= pdal.Writer(filename=output_filename, a_srs=spatial_ref)
+        pipeline |= pdal.Writer(filename=output_filename, forward="all", **params)
 
         logging.info(pipeline.toJSON())
         pipeline.execute()